Former-commit-id: ca286938690a402cd45d1d4dbb366d4161651880

+++ source/plugins/Acquisition/logic/cxUSSavingRecorder.cpp
@@ -159,7 +159,7 @@ void USSavingRecorder::clearRecording()
 }
 
 
-unsigned USSavingRecorder::getNumberOfSavingThreads() const
+size_t USSavingRecorder::getNumberOfSavingThreads() const
 {
 	return mSaveThreads.size();
 }
